requirements.txt
imgui-bundle==1.92.5
opencv-python==4.13.0.92
PyOpenGL==3.1.10
numpy==2.4.2
ultralytics==8.4.16